packages/retrieval/colbert/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retrieval-colbert"
version = "0.1.0"
description = "ColBERT retrieval tooling"
requires-python = ">=3.11"
authors = [{ name = "StratMaster" }]
dependencies = [
  "pydantic>=2.11.0",
  "typer>=0.12.5",
  "PyYAML>=6.0",
]

[project.scripts]
colbert-index = "colbert.index:app"
colbert-search = "colbert.search:app"
colbert-eval = "colbert.eval:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
